refactored from judges to contracts
