+++ Smart-Contract-Verifier-Server/src/main/kotlin/com/savannasolutions/SmartContractVerifierServer/user/models/User.kt
@@ -1,5 +1,5 @@
 package com.savannasolutions.SmartContractVerifierServer.user.models
-import com.savannasolutions.SmartContractVerifierServer.judges.models.Judges
+import com.savannasolutions.SmartContractVerifierServer.contracts.models.Judges
 import com.savannasolutions.SmartContractVerifierServer.messenger.models.MessageStatus
 import com.savannasolutions.SmartContractVerifierServer.messenger.models.Messages
 import com.savannasolutions.SmartContractVerifierServer.negotiation.models.Agreements
