+++ Smart-Contract-Verifier-Server/src/test/kotlin/com/savannasolutions/SmartContractVerifierServer/UnitTests/messenger/MessengerServiceTest.kt
@@ -1,7 +1,7 @@
 package com.savannasolutions.SmartContractVerifierServer.UnitTests.messenger
 
 import com.savannasolutions.SmartContractVerifierServer.common.ResponseStatus
-import com.savannasolutions.SmartContractVerifierServer.judges.repositories.JudgesRepository
+import com.savannasolutions.SmartContractVerifierServer.contracts.repositories.JudgesRepository
 import com.savannasolutions.SmartContractVerifierServer.messenger.models.MessageStatus
 import com.savannasolutions.SmartContractVerifierServer.messenger.models.Messages
 import com.savannasolutions.SmartContractVerifierServer.messenger.repositories.MessageStatusRepository
